Implement aggregator feature
It works on object hierarchy searching for class and it will inject object
in bottom up hierarchy (deepest children injected first). All classes
intended to be scanned for injection has to be marked @Aggregator and all
fields in @Aggregator class intended to be DId have to have
@InjectableField annotated; if beanName with @InjectableField is not
provided then it will use the field name as beanName.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>