Add implementation for an input stream for string
A simple StringReader based input stream for a String. Its added in a
single module which contains simple stand alone utility classes only.

While adding this new module its also ensured that no versions are again
hardcoded in the submodule, but rather they are rather added as properties
to the main POM and then referred to from the submodule POMs.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>