Remove Spring Framework dependency from properties locator
Previously properties locator was depending of Spring for loading
properties file from different types of resources and different encoding.
With this change all those dependency is removed and it solely depends on
Apache Commons Lang module.

In process added a simple Resource API similar to that of spring but
simpler and also added capability to read custom encoded properties file
without adding dependency to JAVA SE 6.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>