Add factory methods to convert class and annotation class names
Class and annotation class names as used in visitor has a different format
in string than that of the '.' separated class names & so added convenient
factory methods to covert class to string and vice versa. Please use them
in proper order and for that you might want to check the test case as well.

The test case was not being used properly and that is also fixed in this
round.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>