Totally separate Spring Framework related implementations
This is done so that the aggregator API and multi location properties
locator can be used in projects without having to import spring.

This change was also inspired by the fact that there is no need of spring
framework dependencies if someone wants to simply mark their domain class
with injection and aggragation annotations and by the fact that a google
guice integration could also be lurking nearby.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>