Refactor abstract generic adapter implementation
The abstract implementation is now divided into concrete implementation and
abstract helper. The refactor is motivated to keep the helper impl and
adapter impl separate; they are typically a has-a relation ship and with
composition more decoupling is achieved.

Add test cases to satisfy the decoupling. Add JMock for test cases.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>