Add context loader for Spring config from classpath directly
The advantage it will add is that it will allow to load from classpath and
thus when combined with BeanFactoryRegistrar it will totally abstract the
DI container from the code and limiting coupling only in configurations.

Add a mock servlet to check for registration of factory. Add test to
integration-test life phase of maven. Use Ant to combine both classes and
target-classes. The integration test does a curl to the servlet to determine
whether it is working or not.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>