Add implementation to read entries of a feed
The implementation concentrates of iterating over the entries of the feed
provided and only that feed. It intends to support reading both nested
content or link. Currently the nested content part is not implemented.

If the link of an entry points to another feed it will search for a link of
the feed to fetch the de-serializable content and will recurse as long as
a feed is found through the link. It assumes that the Jersey client is
configured to de-serialize the MIME Type user have specified in the link
spec.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>