Add capability to read from entry's content
As mentioned earlier we also wanted to be able to read content of an entry
and deserialize it to its entity.

We also did not want to keep any knowledge of how-to deserialize in our
system and thus introduced a stream based entity deserializer who will
know it or use some auto-deserializer.

With this change there are 2 constructors and thus the reader should be
initialized with care.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>