Change project configuration for atom feed reader
The first change is to change the artifact id of the project to atom-client
so that atom server side util can be a separate module

The second change is to minimize the amount of dependency to ensure that
unnecessary dependencies are not present as this modules dependency.

Please note that to get it to run other dependencies might be required by
projects using it.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>