Add HttpClient to ensure that all requested URIs are absolute
If the feed links' have relative URI there is no way for the atom-client to
know how to make it absolute and thus the introduction of HttpClient. It
will keep it simple and have host name and port to convert the URI to an
absolute URI.

Add dependency to commons-lang to get utility functions such as StringUtils

Integrate HttpClient into feed entry reader right away so that it works.


Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>