Add operations to check whether OpenSearch elements are present in a feed
For this first add the dependency to abdera open search description.

Then add a generic operation to check the existence of any XML QName within
the feed and use it to determine whether there is items per page and total
results or not. In the same way add a generic operation to fetch int from
an integer element of OpenSearch Ext. and use it to return total results
and items per page

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>