Add list impl to iterate through a paginated open search result feed
This list will fetch entities dynamically from a open search result feed,
i.e. a feed with open search qnames specified. The specific qname that it
depends on is total result.

This list implementation is in its rudimentary form in terms of
intelligence and thus assumes that it is at the root page of the result.
Thus it only goes "next" blindly. If index is not found it will return null
object.

Later upgrade to this implementation should include sensing page index of
this page and to front or back as required.

Add getter methods for wrapper members to be used in the list to check
for open search elements.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>