Add list impl to load a paginated feed at once in background
It will basically load all entities until there is no such loadable entity
or the max has been reached/crossed in the last lookup.

The loading is done in a background thread. Any call to size() or get(int)
will result in the thread to wait until the background thread is done with
the loading.


Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>