Add Jetty based test structure for client test
The client test will use jetty in an embedded manner to test the functions
of the module.

In process add Jetty dependencies. Use Junit 4 for testing in order to get
BeforeClass and AfterClass. Start and stop the server before any test case
in the class is invoked. Prepare 2 dummy test cases to ensure that config
works.

The atom and json dependencies are for future. Next step would be to get a
Jersey based RESTful WS server running.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>