Fix Url rel API as per specification
The specification says that there can be multiple rels, space delimited, &
the meaning of all the rels are beyond the specification. Permissible rels
are alphanumeric lowercase string tokens and valid URLs. The specification
defines four tokens with defined meanings.

In accordance, the tokens with defined meanings are RelEnum. Url can have a
collection of Rel and if none specified then results is its default. Rel
can be any lowercase alphanumeric string token or URL. The builder supports
both adding and removing Rel.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>