Add Open Search Descriptor XML Serialization implementation
The implementation is coupled with its basic test cases and it seems to
work fine.

Currently for custom attribute values of role of query and other similar
cases minimum 2 customAttributes needs to be added
  1. xmlns:prefix - URI attribute
  2. prefix:dummy - dummyValue attribute to get the attribute be written.
This problem is already being discussed in the xom-interest mailing list.
As soon as a resolution is reached it will be incorporated back into this
implementation.

All elements and attribute names of OpenSearchDescriptor is in XmlConstants
interface and are prefixed ELEM and ATTR respectively.

Deleted the dummy test case file.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>