Fix only adding a namespace attribute for extension of query role
Previously there was a bug due to deficiency of knowledge of XOM but help
from Michael Kay <mike@saxonica.com> of XOM-interest mailing list found out
the right API method and the result is that only namespace can be added.

Add a test case to test both the possible uses of namespace.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>