Add DOM parser for open search descriptor specification 1.1 Draft 4
A test case for the reading is implemented using the test strings used to
test the output of the writer. Match is done by matching original string
vs serialized string of the parsed object from original string.

Changed the test data and modifed indexOffset and pageOffset in one of the
Url's

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>