Add JAX-RS reader provider for open search descriptor
A test case is written to ensure that it does read the descriptor properly
to verify it the the domain serialized again to check the string form to
that of what was expected.

Add a service for the message body reader so that the provider is picked if
meta-inf services is supported.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>