Users invoking POST, PUT and/or DELETE can now check for expected status
It would be a common step for clients to check for Status they expect, thus
introduced them to the WritableResource API and implemented them in the
abstract class. As a result clients will not have to write code to check
status. If no status is provided it will simply ignore checking it.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>