Add setting on bean name in bean factory
The setting will be used to allow InjectableField's name as the sole source
of name of the bean, i.e. it will allow empty bean name, if and only if the
bean factory supports blank bean name.

This change is specially introduced for Google Guice so that users are not
constrained into using @Name annotation.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>