Add utility for loading modules and context name from properties file
First initializing the util using one of getInstance methods and then
invoking register will register an injector with BeanFactorRegistrar. To be
noted is that the Util can be totally by passed and it could be done
manually but this could be useful to abstract the module class names.

Next step would be to test out Google Guice with Smart Bean Util.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>