Incorporate changes from Jersey master
The changes are related to incorporation of the patch after some refactors.
One of the refactors needs to move ThreadLocal based access from jersey
code to our code, making perfect sense.

The change that needs further discussion in Jersey is the preemptive auth
and timeout setting, now these settings needs to be copied whoever wants to
use apache client as a base.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>